Utilities and tools package

Package timeSeries is part of the Rmetrics suit of R packages. It provides a class, timeSeries, particularly aimed at analysis of financial data, along with many methods, functions, and utilities for statistical and financial computations on time series.

Author

Diethelm Wuertz [aut] (original code), Tobias Setz [aut], Yohan Chalabi [aut], Martin Maechler [ctb] (<https://orcid.org/0000-0002-8685-9910>), Georgi N. Boshnakov [cre, ctb]

Maintainer: Georgi N. Boshnakov <georgi.boshnakov@manchester.ac.uk>

timeSeries - S4 timeSeries Class

timeSeries	Creates a 'timeSeries' from scratch

getDataPart, series	...

getUnits	Extracts the time serie units

getTime, time	Extracts the positions of timestamps

use: slot	Extracts the format of the timestamp

getFinCenter, finCenter	Extracts the financial center

use: slot	Extracts the record IDs

getTitle	Extracts the title

use: slot	Extracts the documentation

Base Time Series Functions

apply	Applies a function to blocks of a 'timeSeries'

attach	Attaches a 'timeSeries' to the search path

cbind	Combines columns of two 'timeSeries' objects

rbind	Combines rows of two 'timeSeries' objects

diff	Returns differences of a 'timeSeries' object

dim	returns dimensions of a 'timeSeries' object

merge	Merges two 'timeSeries' objects

rank	Returns sample ranks of a 'timeSeries' object

rev	Reverts a 'timeSeries' object

sample	Resamples a 'timeSeries' object

scale	Scales a 'timeSeries' object

sort	Sorts a 'timeSeries' object

start	Returns start date/time of a 'timeSeries'

end	Returns end date/time of a 'timeSeries'

t	Returns the transpose of a 'timeSeries' object

Subsetting 'timeSeries' Objects

.subset_	Subsets 'timeSeries' objects

.findIndex	Index search in a 'timeSeries' object

[	Subsets a 'timeSeries' object

[<-]	Assigns values to a subset

$	Subsets a 'timeSeries' by column names

$<-	Replaces Subset by column names

t	Returns the transpose of a 'timeSeries'

head	Returns the head of a 'timeSeries'

tail	Returns the tail of a time Series

na.omit	Handles NAs in a timeSeries object

removeNA	removes NAs from a matrix object

substituteNA	substitutes NAs by zero, column mean or median

interpNA	interpolates NAs using R's "approx" function

Mathematical Operation

Ops.timeSeries	S4: Arith method for a 'timeSeries' object

abs	Returns absolute values of a 'timeSeries' object

sqrt	Returns square root of a 'timeSeries' object

exp	Returns the exponential values of a 'timeSeries' object

log	Returns the logarithm of a 'timeSeries' object

sign	Returns the signs of a 'timeSeries' object

diff	Differences a 'timeSeries' object

scale	Centers and/or scales a 'timeSeries' object

quantile	Returns quantiles of an univariate 'timeSeries'

Methods

as.timeSeries	Defines method for a 'timeSeries'

as.*.default	Returns the input

as.*.ts	Transforma a 'ts' object into a 'timeSeries'

as.*.data.frame	Transforms a 'data.frame' intp a 'timeSeries

as.*.character	Loads and transforms from a demo file

as.*.zoo	Transforms a 'zoo' object into a 'timeSeries'

as.vector.*	Converts univariate timeSeries to vector

as.matrix.*	Converts timeSeries to matrix

as.numeric.*	Converts timeSeries to numeric

as.data.frame.*	Converts timeSeries to data.frame

as.ts.*	Converts timeSeries to ts

as.logical.*	Converts timeSeries to logical

is.timeSeries	Tests for a 'timeSeries' object

plot	Displays a X-Y 'timeSeries' Plot

lines	Adds connected line segments to a plot

points	Adds Points to a plot

show	Prints a 'timeSeries oobject

Financial time series functions

align	Aligns a 'timeSeries' to time stamps

cumulated	Computes cumulated series from a returns

alignDailySeries	Aligns a 'timeSeries' to calendarical dates

rollDailySeries	Rolls a 'timeSeries daily

drawdowns	Computes series of drawdowns from financial returns

drawdownsStats	Computes drawdowns statistics

durations	Computes durations from a financial time series

countMonthlyRecords	Counts monthly records in a 'timeSeries'

rollMonthlyWindows	Rolls Monthly windows

rollMonthlySeries	Rolls a 'timeSeries' monthly

endOfPeriodSeries	Returns end of periodical series

endOfPeriodStats	Returns end of period statistics

endOfPeriodBenchmarks	Returns period benchmarks

returns	Computes returns from prices or indexes

returns0	Computes untrimmed returns from prices or indexes

runlengths	Computes run lenghts of a 'timeSeries'

smooth	Smoothes a 'timeSeries'

splits	Detects 'timeSeries' splits by outlier detection

spreads	Computes spreads from a price/index stream

turns	Computes turning points in a 'timeSeries' object

turnsStats	Computes turning points statistics

Statistics Time Series functions

colCumsums	Computes cumulated column sums of a 'timeSeries'

colCummaxs	Computes cumulated maximum of a 'timeSeries'

colCummins	Computes cumulated minimum of a 'timeSeries'

colCumprods	Computes cumulated pruduct values by column

colCumreturns	Computes cumulated returns by column

colSums	Computes sums of all values in each column

colMeans	Computes means of all values in each column

colSds	Computes standard deviations of all values in each column

colVars	Computes variances of all values in each column

colSkewness	Computes skewness of all values in each column

colKurtosis	Computes kurtosis of all values in each column

colMaxs	Computes maxima of all values in each column

colMins	Computes minima of all values in each column

colProds	Computes products of all values in each column

colStats	Computes statistics of all values in each column

orderColnames	Returns ordered column names of a 'timeSeries'

sortColnames	Returns alphabetically sorted column names

sampleColnames	Returns sampled column names of a 'timeSeries'

pcaColnames	Returns PCA correlation ordered column names

hclustColnames	Returns hierarchically clustered columnames

statsColnames	Returns statisticall rearrange columnames

orderStatistics	Computes order statistics of a 'timeSeries' object

rollMean	Computes rolling means of a 'timeSeries' object

rollMin	Computes rolling minima of a 'timeSeries' object

rollMax	Computes rolling maxima of a 'timeSeries' object

rollMedian	Computes rolling medians of a 'timeSeries' object

rollStats	Computes rolling statistics of a 'timeSeries' objectcr

rowCumsums	Computes cumulated column sums of a 'timeSeries'

smoothLowess	Smoothes a series with lowess function

smoothSupsmu	Smoothes a series with supsmu function

smoothSpline	Smoothes a series with smooth.spline function

Misc Functions

dummyDailySeries	Creates a dummy daily 'timeSeries' object

isMonthly	Decides if the series consists of monthly records

getArgs	Extracts arguments from a S4 method