Subsetting time series

Subset a "timeSeries" objects due to different aspects.

[	"[" method for a "timeSeries" object,

[<-	"[<-" method to assign value for a subset of a "timeSeries" object,

head	Returns the head of a "timeSeries" object,

tail	Returns the tail of a "timeSeries" object,

outliers	Removes outliers from a "timeSeries" object.

Usage

<!-- % \method{[}{timeSeries}(x, i, j, drop) -->
<!-- % \method{[}{timeSeries}(x, i, j) &lt;- value -->

# S4 method for timeSeries
head(x, n = 6, recordIDs = FALSE, ...)
# S4 method for timeSeries
tail(x, n = 6, recordIDs = FALSE, ...)

# S4 method for timeSeries
outlier(x, sd = 3, complement = TRUE, ...)

Arguments

complement

[outlierSeries] -
a logical flag, should the outler series or its complement be returns, by default TRUE which returns the series free of outliers.

n

[head][tail] -
an integer specifying the number of lines to be returned. By default n=6.

recordIDs

[head][tail] -
a logical value. Should the recordIDs returned together with the data matrix and time series positions?

sd

[outlierSeries] -
a numeric value of standard deviations, e.g. 10 means that values larger or smaller tahn ten times the standard deviation will be removed from the series.

x

an object of class timeSeries.

...

arguments passed to other methods.

Value

All functions return an object of class "timeSeries".

Examples

## Create an Artificial 'timeSeries' Object - 
   setRmetricsOptions(myFinCenter = "GMT")
   charvec <- timeCalendar()
   set.seed(4711)
   data <- matrix(exp(cumsum(rnorm(12, sd = 0.1))))
   tS <- timeSeries(data, charvec, units = "tS")
   tS
#> GMT
#>                   tS
#> 2022-01-01 1.1995824
#> 2022-02-01 1.3757753
#> 2022-03-01 1.5506114
#> 2022-04-01 1.4887865
#> 2022-05-01 1.4005479
#> 2022-06-01 1.2043889
#> 2022-07-01 1.3069906
#> 2022-08-01 1.1867998
#> 2022-09-01 1.1815277
#> 2022-10-01 1.2389245
#> 2022-11-01 1.1230263
#> 2022-12-01 0.9596526
   
## Subset Series by Counts "[" - 
   tS[1:3, ]
#> GMT
#>                  tS
#> 2022-01-01 1.199582
#> 2022-02-01 1.375775
#> 2022-03-01 1.550611
       
## Subset the Head of the Series - 
   head(tS, 6)
#> GMT
#>                  tS
#> 2022-01-01 1.199582
#> 2022-02-01 1.375775
#> 2022-03-01 1.550611
#> 2022-04-01 1.488787
#> 2022-05-01 1.400548
#> 2022-06-01 1.204389