Special daily time series

Special daily 'timeSeries' functions.

Usage

alignDailySeries(x, method = c("before", "after", "interp", "fillNA",
    "fmm", "periodic", "natural", "monoH.FC"),
    include.weekends = FALSE, units = NULL, zone = "",
    FinCenter = "", ...)

rollDailySeries(x, period = "7d", FUN, ...)

Arguments

x

an object of class timeSeries.

method

the method to be used for the alignment. A character string, one of "before", use the data from the row whose position is just before the unmatched position, or "after", use the data from the row whose position is just after the unmatched position, or "linear", interpolate linearly between "before" and "after".

include.weekends

a logical value. Should weekend dates be included or removed from the series?

units

an optional character string, which allows to overwrite the current column names of a timeSeries object. By default NULL which means that the column names are selected automatically.

zone

the time zone or financial center where the data were recorded.

FinCenter

a character with the the location of the financial center named as "continent/city".

period

a character string specifying the rollling period composed by the length of the period and its unit, e.g. "7d" represents one week.

FUN

a function to use for aggregation, by default colMeans.

...

arguments passed to interpolating methods.

Details

alignDailySeries aligns a daily 'timeSeries' to new positions,

rollDailySeries rolls daily a 'timeSeries' on a given period.

Value

for alignDailySeries, a weekly aligned daily

timeSeries object from a daily time series with missing holidays.

for rollDailySeries, an object of class timeSeries

with rolling values, computed from the function FUN.

Examples

## Use Microsofts' OHLCV Price Series -
   head(MSFT)
#> GMT
#>               Open    High     Low   Close   Volume
#> 2000-09-27 63.4375 63.5625 59.8125 60.6250 53077800
#> 2000-09-28 60.8125 61.8750 60.6250 61.3125 26180200
#> 2000-09-29 61.0000 61.3125 58.6250 60.3125 37026800
#> 2000-10-02 60.5000 60.8125 58.2500 59.1250 29281200
#> 2000-10-03 59.5625 59.8125 56.5000 56.5625 42687000
#> 2000-10-04 56.3750 56.5625 54.5000 55.4375 68226700
   end(MSFT)
#> GMT
#> [1] [2001-09-27]

## Cut out April Data from 2001 -
   Close <- MSFT[, "Close"]
   tsApril01 <- window(Close, start="2001-04-01", end="2001-04-30")
   tsApril01
#> GMT
#>              Close
#> 2001-04-02 55.8125
#> 2001-04-03 53.3750
#> 2001-04-04 51.9375
#> 2001-04-05 56.7500
#> 2001-04-06 56.1875
#> 2001-04-09 57.1500
#> 2001-04-10 59.6800
#> 2001-04-11 60.0400
#> 2001-04-12 62.1800
#> 2001-04-16 60.7900
#> 2001-04-17 61.4800
#> 2001-04-18 65.4300
#> 2001-04-19 68.0400
#> 2001-04-20 69.0000
#> 2001-04-23 68.2500
#> 2001-04-24 67.5500
#> 2001-04-25 69.6900
#> 2001-04-26 69.1300
#> 2001-04-27 67.1200
#> 2001-04-30 67.7500

## Align Daily Series with NA -
   tsRet <- returns(tsApril01, trim = TRUE)
   GoodFriday(2001)
#> GMT
#> [1] [2001-04-13]
   EasterMonday(2001)
#> GMT
#> [1] [2001-04-16]
   alignDailySeries(tsRet, method = "fillNA", include.weekends = FALSE)
#> GMT
#>                   Close
#> 2001-04-03 -0.044655387
#> 2001-04-04 -0.027301399
#> 2001-04-05  0.088614584
#> 2001-04-06 -0.009961344
#> 2001-04-09  0.016985078
#> 2001-04-10  0.043317566
#> 2001-04-11  0.006014051
#> 2001-04-12  0.035022398
#> 2001-04-13           NA
#> 2001-04-16 -0.022608103
#> 2001-04-17  0.011286617
#> 2001-04-18  0.062268950
#> 2001-04-19  0.039114899
#> 2001-04-20  0.014010737
#> 2001-04-23 -0.010929071
#> 2001-04-24 -0.010309370
#> 2001-04-25  0.031188771
#> 2001-04-26 -0.008068045
#> 2001-04-27 -0.029506728
#> 2001-04-30  0.009342398
   alignDailySeries(tsRet, method = "fillNA", include.weekends = TRUE)
#> GMT
#>                   Close
#> 2001-04-03 -0.044655387
#> 2001-04-04 -0.027301399
#> 2001-04-05  0.088614584
#> 2001-04-06 -0.009961344
#> 2001-04-07           NA
#> 2001-04-08           NA
#> 2001-04-09  0.016985078
#> 2001-04-10  0.043317566
#> 2001-04-11  0.006014051
#> 2001-04-12  0.035022398
#> 2001-04-13           NA
#> 2001-04-14           NA
#> 2001-04-15           NA
#> 2001-04-16 -0.022608103
#> 2001-04-17  0.011286617
#> 2001-04-18  0.062268950
#> 2001-04-19  0.039114899
#> 2001-04-20  0.014010737
#> 2001-04-21           NA
#> 2001-04-22           NA
#> 2001-04-23 -0.010929071
#> 2001-04-24 -0.010309370
#> 2001-04-25  0.031188771
#> 2001-04-26 -0.008068045
#> 2001-04-27 -0.029506728
#> 2001-04-28           NA
#> 2001-04-29           NA
#> 2001-04-30  0.009342398

## Align Daily Series by Interpolated Values -
   alignDailySeries(tsRet, method = "interp", include.weekend = FALSE)
#> GMT
#>                   Close
#> 2001-04-03 -0.044655387
#> 2001-04-04 -0.027301399
#> 2001-04-05  0.088614584
#> 2001-04-06 -0.009961344
#> 2001-04-09  0.016985078
#> 2001-04-10  0.043317566
#> 2001-04-11  0.006014051
#> 2001-04-12  0.035022398
#> 2001-04-13  0.020614773
#> 2001-04-16 -0.022608103
#> 2001-04-17  0.011286617
#> 2001-04-18  0.062268950
#> 2001-04-19  0.039114899
#> 2001-04-20  0.014010737
#> 2001-04-23 -0.010929071
#> 2001-04-24 -0.010309370
#> 2001-04-25  0.031188771
#> 2001-04-26 -0.008068045
#> 2001-04-27 -0.029506728
#> 2001-04-30  0.009342398
   alignDailySeries(tsRet, method = "interp", include.weekend = TRUE)
#> GMT
#>                    Close
#> 2001-04-03 -0.0446553871
#> 2001-04-04 -0.0273013989
#> 2001-04-05  0.0886145837
#> 2001-04-06 -0.0099613441
#> 2001-04-07 -0.0009792034
#> 2001-04-08  0.0080029373
#> 2001-04-09  0.0169850780
#> 2001-04-10  0.0433175657
#> 2001-04-11  0.0060140509
#> 2001-04-12  0.0350223979
#> 2001-04-13  0.0206147727
#> 2001-04-14  0.0062071475
#> 2001-04-15 -0.0082004777
#> 2001-04-16 -0.0226081029
#> 2001-04-17  0.0112866169
#> 2001-04-18  0.0622689502
#> 2001-04-19  0.0391148986
#> 2001-04-20  0.0140107371
#> 2001-04-21  0.0056974679
#> 2001-04-22 -0.0026158013
#> 2001-04-23 -0.0109290705
#> 2001-04-24 -0.0103093697
#> 2001-04-25  0.0311887710
#> 2001-04-26 -0.0080680455
#> 2001-04-27 -0.0295067278
#> 2001-04-28 -0.0165570193
#> 2001-04-29 -0.0036073109
#> 2001-04-30  0.0093423976