Model Frames for Time Series Objects

Allow to work with model frames for 'timeSeries' objects.

Value

Returns an object of class 'timeSeries.

Details

The function model.frame is a generic function which returns in the R-ststs framework by default a data.frame with the variables needed to use formula and any ... arguments. In contrast to this the method returns an object of class timeSeries when the argument data was not a data.frame but also an object of class 'timeSeries'.

Note

This function is preliminary and untested.

See also

Examples

## Load Microsoft Data -
   setRmetricsOptions(myFinCenter = "GMT")
   X <- MSFT[1:12, ]

## Extract High's and Low's:
   DATA <- model.frame( ~ High + Low, data = X)
   class(DATA)
#> [1] "data.frame"
   as.timeSeries(DATA)
#> GMT
#>               High     Low
#> 2000-09-27 63.5625 59.8125
#> 2000-09-28 61.8750 60.6250
#> 2000-09-29 61.3125 58.6250
#> 2000-10-02 60.8125 58.2500
#> 2000-10-03 59.8125 56.5000
#> 2000-10-04 56.5625 54.5000
#> 2000-10-05 57.2500 55.2500
#> 2000-10-06 56.7500 54.7500
#> 2000-10-09 55.7500 53.0000
#> 2000-10-10 55.5625 53.8125
#> 2000-10-11 56.9375 54.0000
#> 2000-10-12 56.8750 53.8125
   
## Extract Open Prices and their log10's:
   base <- 10
   Open <- model.frame(Open ~ log(Open, base = `base`), data = X)
   colnames(Open) <- c("X", "log10(X)")
   class(Open)
#> [1] "data.frame"
   as.timeSeries(Open)
#> GMT
#>                  X log10(X)
#> 2000-09-27 63.4375 1.802346
#> 2000-09-28 60.8125 1.783993
#> 2000-09-29 61.0000 1.785330
#> 2000-10-02 60.5000 1.781755
#> 2000-10-03 59.5625 1.774973
#> 2000-10-04 56.3750 1.751087
#> 2000-10-05 55.5000 1.744293
#> 2000-10-06 55.8125 1.746731
#> 2000-10-09 55.6250 1.745270
#> 2000-10-10 53.9375 1.731891
#> 2000-10-11 54.0000 1.732394
#> 2000-10-12 56.3125 1.750605