Handle missing values in 'timeSeries' objects

Functions for handling missing values in "timeSeries" objects.

Usage

# S3 method for class 'timeSeries'
na.omit(object, method = c("r", "s", "z", "ir", "iz", "ie"), 
    interp = c("before", "linear", "after"), FUN, ...)

Arguments

object

an object of class "timeSeries".

method

the method of handling NAs, see section ‘Details’.

interp

Three alternative methods are provided to remove NAs from the data: type="zeros" replaces the missing values with zeros, type="mean" replaces the missing values with the column mean, type="median" replaces the missing values with the column median.

FUN

a function or a name of a function, such as "mean" or median. FUN is applied to the non-NA values in each column to determine the replacement value. The call looks like FUN(coli, na.rm = TRUE), so FUN should have argument na.rm. All arguments except object are ignored if FUN is specified.

...

arguments to be passed to the function as.matrix.

Details

Functions for handling missing values in "timeSeries" objects and in objects which can be transformed into a vector or a two dimensional matrix.

For na.omit argument method specifies how to handle NAs. Can be one of the following strings:

method = "s"

na.rm = FALSE, skip, i.e. do nothing,

method = "r"

remove NAs,

method = "z"

substitute NAs by zeros,

method = "ir"

interpolate NAs and remove NAs at the beginning and end of the series,

method = "iz"

interpolate NAs and substitute NAs at the beginning and end of the series,

method = "ie"

interpolate NAs and extrapolate NAs at the beginning and end of theseries.

Note

When dealing with daily data sets, there exists another function alignDailySeries which can handle missing data in un-aligned calendrical "timeSeries" objects.

Additional remarks by GNB:

removeNA(x) is equivalent to na.omit(x) or na.omit(x), methods = "r".

interpNA can be replaced by a call to na.omit with argument method equal to ir, iz, or ie, and argument "interp" equal to the "method" argument for interpNA (note that the defaults are not the same).

substituteNA(x, type = "zeros") is equivalent to na.omit(x, method = "z"). For other values of type one can use argument FUN, as in na.omit(x, FUN = "mean").

A final remark: the three deprecated functions are non-generic. removeNA(x) is completely redundant as it simply calls na.omit. The other two however may be useful for matrix-like objects. Please inform the maintainer of the package if you use them on objects other than from class "timeSeries" and wish them kept in the future.

References

Troyanskaya O., Cantor M., Sherlock G., Brown P., Hastie T., Tibshirani R., Botstein D., Altman R.B., (2001); Missing Value Estimation Methods for DNA microarrays Bioinformatics 17, 520–525.

See also

Examples

X <- matrix(rnorm(100), ncol = 5)  # Create a Matrix X
X[3, 5] <- NA                      # Replace a Single NA Inside
X[17, 2:4] <- c(NA, NA, NA)        # Replace Three in a Row Inside
X[13:15, 4] <- c(NA, NA, NA)       # Replace Three in a Column Inside
X[11:12, 5] <- c(NA, NA)           # Replace Two at the Right Border
X[20, 1] <- NA                     # Replace One in the Lower Left Corner
X
#>             [,1]        [,2]        [,3]        [,4]       [,5]
#>  [1,] -1.2900417 -0.43898655 -0.66582924 -0.74490561 -1.1151520
#>  [2,]  0.4222867 -1.32640195  1.31724791 -1.50875006  1.4217720
#>  [3,] -0.1030881 -1.19271949  0.13421979 -0.95380354         NA
#>  [4,]  0.5258783 -1.13269775  0.33373548  0.13156962 -1.2167619
#>  [5,]  0.4992021 -0.71374675  1.42513695 -0.10488850 -1.6879300
#>  [6,]  1.2296179  0.97129225 -0.66687363 -1.29914179 -0.8430539
#>  [7,]  0.4359482  0.11081679 -0.15419999 -1.81072734  1.3052824
#>  [8,] -0.7221102  1.14315208  0.39575880  0.34617192  0.2354969
#>  [9,]  0.8468178 -0.79418025 -0.28903724  0.30310787  0.7766441
#> [10,]  1.9957190 -0.09344266 -1.03946394  0.62199075 -0.8366753
#> [11,]  0.8602705 -0.04080370  0.91980757 -0.07572521         NA
#> [12,] -0.1299832 -0.76676070  1.18297818  0.68143284         NA
#> [13,] -0.2294296  2.04819505 -0.06661768          NA  0.7639455
#> [14,]  0.1092238 -0.70144007  0.69121516          NA  0.1272981
#> [15,]  0.7522100  0.65586854 -1.28738212          NA -2.3297949
#> [16,]  0.9254184  0.45330809  0.89816897 -0.79592837 -0.9856772
#> [17,] -0.2917223          NA          NA          NA  0.4704501
#> [18,] -0.4709085  0.66014159 -0.53506265  0.07207710  2.2224636
#> [19,]  0.1615685  0.37914099 -1.39575652  1.05144407 -0.1124329
#> [20,]         NA -0.18022159 -0.58244338  1.17451229 -0.1805846
Xts <- timeSeries(X)  # convert X to timeSeries Xts

## remove rows with NAs
na.omit(Xts)
#>  
#>             SS.1        SS.2       SS.3       SS.4       SS.5
#>  [1,] -1.2900417 -0.43898655 -0.6658292 -0.7449056 -1.1151520
#>  [2,]  0.4222867 -1.32640195  1.3172479 -1.5087501  1.4217720
#>  [3,]  0.5258783 -1.13269775  0.3337355  0.1315696 -1.2167619
#>  [4,]  0.4992021 -0.71374675  1.4251369 -0.1048885 -1.6879300
#>  [5,]  1.2296179  0.97129225 -0.6668736 -1.2991418 -0.8430539
#>  [6,]  0.4359482  0.11081679 -0.1542000 -1.8107273  1.3052824
#>  [7,] -0.7221102  1.14315208  0.3957588  0.3461719  0.2354969
#>  [8,]  0.8468178 -0.79418025 -0.2890372  0.3031079  0.7766441
#>  [9,]  1.9957190 -0.09344266 -1.0394639  0.6219907 -0.8366753
#> [10,]  0.9254184  0.45330809  0.8981690 -0.7959284 -0.9856772
#> [11,] -0.4709085  0.66014159 -0.5350626  0.0720771  2.2224636
#> [12,]  0.1615685  0.37914099 -1.3957565  1.0514441 -0.1124329

## Subsitute NA's with zeros or column means (formerly substituteNA())
na.omit(Xts, method = "z")
#>  
#>             SS.1        SS.2        SS.3        SS.4       SS.5
#>  [1,] -1.2900417 -0.43898655 -0.66582924 -0.74490561 -1.1151520
#>  [2,]  0.4222867 -1.32640195  1.31724791 -1.50875006  1.4217720
#>  [3,] -0.1030881 -1.19271949  0.13421979 -0.95380354  0.0000000
#>  [4,]  0.5258783 -1.13269775  0.33373548  0.13156962 -1.2167619
#>  [5,]  0.4992021 -0.71374675  1.42513695 -0.10488850 -1.6879300
#>  [6,]  1.2296179  0.97129225 -0.66687363 -1.29914179 -0.8430539
#>  [7,]  0.4359482  0.11081679 -0.15419999 -1.81072734  1.3052824
#>  [8,] -0.7221102  1.14315208  0.39575880  0.34617192  0.2354969
#>  [9,]  0.8468178 -0.79418025 -0.28903724  0.30310787  0.7766441
#> [10,]  1.9957190 -0.09344266 -1.03946394  0.62199075 -0.8366753
#> [11,]  0.8602705 -0.04080370  0.91980757 -0.07572521  0.0000000
#> [12,] -0.1299832 -0.76676070  1.18297818  0.68143284  0.0000000
#> [13,] -0.2294296  2.04819505 -0.06661768  0.00000000  0.7639455
#> [14,]  0.1092238 -0.70144007  0.69121516  0.00000000  0.1272981
#> [15,]  0.7522100  0.65586854 -1.28738212  0.00000000 -2.3297949
#> [16,]  0.9254184  0.45330809  0.89816897 -0.79592837 -0.9856772
#> [17,] -0.2917223  0.00000000  0.00000000  0.00000000  0.4704501
#> [18,] -0.4709085  0.66014159 -0.53506265  0.07207710  2.2224636
#> [19,]  0.1615685  0.37914099 -1.39575652  1.05144407 -0.1124329
#> [20,]  0.0000000 -0.18022159 -0.58244338  1.17451229 -0.1805846
na.omit(Xts, FUN = "mean")
#>  
#>             SS.1        SS.2        SS.3        SS.4       SS.5
#>  [1,] -1.2900417 -0.43898655 -0.66582924 -0.74490561 -1.1151520
#>  [2,]  0.4222867 -1.32640195  1.31724791 -1.50875006  1.4217720
#>  [3,] -0.1030881 -1.19271949  0.13421979 -0.95380354 -0.1167476
#>  [4,]  0.5258783 -1.13269775  0.33373548  0.13156962 -1.2167619
#>  [5,]  0.4992021 -0.71374675  1.42513695 -0.10488850 -1.6879300
#>  [6,]  1.2296179  0.97129225 -0.66687363 -1.29914179 -0.8430539
#>  [7,]  0.4359482  0.11081679 -0.15419999 -1.81072734  1.3052824
#>  [8,] -0.7221102  1.14315208  0.39575880  0.34617192  0.2354969
#>  [9,]  0.8468178 -0.79418025 -0.28903724  0.30310787  0.7766441
#> [10,]  1.9957190 -0.09344266 -1.03946394  0.62199075 -0.8366753
#> [11,]  0.8602705 -0.04080370  0.91980757 -0.07572521 -0.1167476
#> [12,] -0.1299832 -0.76676070  1.18297818  0.68143284 -0.1167476
#> [13,] -0.2294296  2.04819505 -0.06661768 -0.18197275  0.7639455
#> [14,]  0.1092238 -0.70144007  0.69121516 -0.18197275  0.1272981
#> [15,]  0.7522100  0.65586854 -1.28738212 -0.18197275 -2.3297949
#> [16,]  0.9254184  0.45330809  0.89816897 -0.79592837 -0.9856772
#> [17,] -0.2917223 -0.05049927  0.03240013 -0.18197275  0.4704501
#> [18,] -0.4709085  0.66014159 -0.53506265  0.07207710  2.2224636
#> [19,]  0.1615685  0.37914099 -1.39575652  1.05144407 -0.1124329
#> [20,]  0.2908883 -0.18022159 -0.58244338  1.17451229 -0.1805846
na.omit(Xts, FUN = "median")
#>  
#>             SS.1        SS.2        SS.3         SS.4       SS.5
#>  [1,] -1.2900417 -0.43898655 -0.66582924 -0.744905605 -1.1151520
#>  [2,]  0.4222867 -1.32640195  1.31724791 -1.508750056  1.4217720
#>  [3,] -0.1030881 -1.19271949  0.13421979 -0.953803536 -0.1124329
#>  [4,]  0.5258783 -1.13269775  0.33373548  0.131569622 -1.2167619
#>  [5,]  0.4992021 -0.71374675  1.42513695 -0.104888498 -1.6879300
#>  [6,]  1.2296179  0.97129225 -0.66687363 -1.299141789 -0.8430539
#>  [7,]  0.4359482  0.11081679 -0.15419999 -1.810727344  1.3052824
#>  [8,] -0.7221102  1.14315208  0.39575880  0.346171923  0.2354969
#>  [9,]  0.8468178 -0.79418025 -0.28903724  0.303107872  0.7766441
#> [10,]  1.9957190 -0.09344266 -1.03946394  0.621990749 -0.8366753
#> [11,]  0.8602705 -0.04080370  0.91980757 -0.075725215 -0.1124329
#> [12,] -0.1299832 -0.76676070  1.18297818  0.681432842 -0.1124329
#> [13,] -0.2294296  2.04819505 -0.06661768 -0.001824058  0.7639455
#> [14,]  0.1092238 -0.70144007  0.69121516 -0.001824058  0.1272981
#> [15,]  0.7522100  0.65586854 -1.28738212 -0.001824058 -2.3297949
#> [16,]  0.9254184  0.45330809  0.89816897 -0.795928368 -0.9856772
#> [17,] -0.2917223 -0.09344266 -0.06661768 -0.001824058  0.4704501
#> [18,] -0.4709085  0.66014159 -0.53506265  0.072077100  2.2224636
#> [19,]  0.1615685  0.37914099 -1.39575652  1.051444068 -0.1124329
#> [20,]  0.4222867 -0.18022159 -0.58244338  1.174512287 -0.1805846

## Subsitute NA's with a trimmed mean
na.omit(Xts, FUN = function(x, na.rm) mean(x, trim = 0.10, na.rm = na.rm))
#>  
#>             SS.1        SS.2        SS.3        SS.4       SS.5
#>  [1,] -1.2900417 -0.43898655 -0.66582924 -0.74490561 -1.1151520
#>  [2,]  0.4222867 -1.32640195  1.31724791 -1.50875006  1.4217720
#>  [3,] -0.1030881 -1.19271949  0.13421979 -0.95380354 -0.1251586
#>  [4,]  0.5258783 -1.13269775  0.33373548  0.13156962 -1.2167619
#>  [5,]  0.4992021 -0.71374675  1.42513695 -0.10488850 -1.6879300
#>  [6,]  1.2296179  0.97129225 -0.66687363 -1.29914179 -0.8430539
#>  [7,]  0.4359482  0.11081679 -0.15419999 -1.81072734  1.3052824
#>  [8,] -0.7221102  1.14315208  0.39575880  0.34617192  0.2354969
#>  [9,]  0.8468178 -0.79418025 -0.28903724  0.30310787  0.7766441
#> [10,]  1.9957190 -0.09344266 -1.03946394  0.62199075 -0.8366753
#> [11,]  0.8602705 -0.04080370  0.91980757 -0.07572521 -0.1251586
#> [12,] -0.1299832 -0.76676070  1.18297818  0.68143284 -0.1251586
#> [13,] -0.2294296  2.04819505 -0.06661768 -0.16252492  0.7639455
#> [14,]  0.1092238 -0.70144007  0.69121516 -0.16252492  0.1272981
#> [15,]  0.7522100  0.65586854 -1.28738212 -0.16252492 -2.3297949
#> [16,]  0.9254184  0.45330809  0.89816897 -0.79592837 -0.9856772
#> [17,] -0.2917223 -0.09889878  0.03448365 -0.16252492  0.4704501
#> [18,] -0.4709085  0.66014159 -0.53506265  0.07207710  2.2224636
#> [19,]  0.1615685  0.37914099 -1.39575652  1.05144407 -0.1124329
#> [20,]  0.2836000 -0.18022159 -0.58244338  1.17451229 -0.1805846

## interpolate NA's linearily (formerly interpNA())
na.omit(X, method = "ir", interp = "linear")
#>             [,1]        [,2]       [,3]       [,4]       [,5]
#>  [1,] -1.2900417 -0.43898655 -0.6658292 -0.7449056 -1.1151520
#>  [2,]  0.4222867 -1.32640195  1.3172479 -1.5087501  1.4217720
#>  [3,]  0.5258783 -1.13269775  0.3337355  0.1315696 -1.2167619
#>  [4,]  0.4992021 -0.71374675  1.4251369 -0.1048885 -1.6879300
#>  [5,]  1.2296179  0.97129225 -0.6668736 -1.2991418 -0.8430539
#>  [6,]  0.4359482  0.11081679 -0.1542000 -1.8107273  1.3052824
#>  [7,] -0.7221102  1.14315208  0.3957588  0.3461719  0.2354969
#>  [8,]  0.8468178 -0.79418025 -0.2890372  0.3031079  0.7766441
#>  [9,]  1.9957190 -0.09344266 -1.0394639  0.6219907 -0.8366753
#> [10,]  0.9254184  0.45330809  0.8981690 -0.7959284 -0.9856772
#> [11,] -0.4709085  0.66014159 -0.5350626  0.0720771  2.2224636
#> [12,]  0.1615685  0.37914099 -1.3957565  1.0514441 -0.1124329
#> attr(,"na.action")
#> [1] 20 17 13 14 15  3 11 12
#> attr(,"class")
#> [1] "omit"
na.omit(X, method = "iz", interp = "linear")
#>             [,1]        [,2]       [,3]       [,4]       [,5]
#>  [1,] -1.2900417 -0.43898655 -0.6658292 -0.7449056 -1.1151520
#>  [2,]  0.4222867 -1.32640195  1.3172479 -1.5087501  1.4217720
#>  [3,]  0.5258783 -1.13269775  0.3337355  0.1315696 -1.2167619
#>  [4,]  0.4992021 -0.71374675  1.4251369 -0.1048885 -1.6879300
#>  [5,]  1.2296179  0.97129225 -0.6668736 -1.2991418 -0.8430539
#>  [6,]  0.4359482  0.11081679 -0.1542000 -1.8107273  1.3052824
#>  [7,] -0.7221102  1.14315208  0.3957588  0.3461719  0.2354969
#>  [8,]  0.8468178 -0.79418025 -0.2890372  0.3031079  0.7766441
#>  [9,]  1.9957190 -0.09344266 -1.0394639  0.6219907 -0.8366753
#> [10,]  0.9254184  0.45330809  0.8981690 -0.7959284 -0.9856772
#> [11,] -0.4709085  0.66014159 -0.5350626  0.0720771  2.2224636
#> [12,]  0.1615685  0.37914099 -1.3957565  1.0514441 -0.1124329
#> attr(,"na.action")
#> [1] 20 17 13 14 15  3 11 12
#> attr(,"class")
#> [1] "omit"
na.omit(X, method = "ie", interp = "linear")
#>             [,1]        [,2]       [,3]       [,4]       [,5]
#>  [1,] -1.2900417 -0.43898655 -0.6658292 -0.7449056 -1.1151520
#>  [2,]  0.4222867 -1.32640195  1.3172479 -1.5087501  1.4217720
#>  [3,]  0.5258783 -1.13269775  0.3337355  0.1315696 -1.2167619
#>  [4,]  0.4992021 -0.71374675  1.4251369 -0.1048885 -1.6879300
#>  [5,]  1.2296179  0.97129225 -0.6668736 -1.2991418 -0.8430539
#>  [6,]  0.4359482  0.11081679 -0.1542000 -1.8107273  1.3052824
#>  [7,] -0.7221102  1.14315208  0.3957588  0.3461719  0.2354969
#>  [8,]  0.8468178 -0.79418025 -0.2890372  0.3031079  0.7766441
#>  [9,]  1.9957190 -0.09344266 -1.0394639  0.6219907 -0.8366753
#> [10,]  0.9254184  0.45330809  0.8981690 -0.7959284 -0.9856772
#> [11,] -0.4709085  0.66014159 -0.5350626  0.0720771  2.2224636
#> [12,]  0.1615685  0.37914099 -1.3957565  1.0514441 -0.1124329
#> attr(,"na.action")
#> [1] 20 17 13 14 15  3 11 12
#> attr(,"class")
#> [1] "omit"
   
## take previous values in a column
na.omit(X, method = "ir", interp = "before")
#>             [,1]        [,2]       [,3]       [,4]       [,5]
#>  [1,] -1.2900417 -0.43898655 -0.6658292 -0.7449056 -1.1151520
#>  [2,]  0.4222867 -1.32640195  1.3172479 -1.5087501  1.4217720
#>  [3,]  0.5258783 -1.13269775  0.3337355  0.1315696 -1.2167619
#>  [4,]  0.4992021 -0.71374675  1.4251369 -0.1048885 -1.6879300
#>  [5,]  1.2296179  0.97129225 -0.6668736 -1.2991418 -0.8430539
#>  [6,]  0.4359482  0.11081679 -0.1542000 -1.8107273  1.3052824
#>  [7,] -0.7221102  1.14315208  0.3957588  0.3461719  0.2354969
#>  [8,]  0.8468178 -0.79418025 -0.2890372  0.3031079  0.7766441
#>  [9,]  1.9957190 -0.09344266 -1.0394639  0.6219907 -0.8366753
#> [10,]  0.9254184  0.45330809  0.8981690 -0.7959284 -0.9856772
#> [11,] -0.4709085  0.66014159 -0.5350626  0.0720771  2.2224636
#> [12,]  0.1615685  0.37914099 -1.3957565  1.0514441 -0.1124329
#> attr(,"na.action")
#> [1] 20 17 13 14 15  3 11 12
#> attr(,"class")
#> [1] "omit"
na.omit(X, method = "iz", interp = "before")
#>             [,1]        [,2]       [,3]       [,4]       [,5]
#>  [1,] -1.2900417 -0.43898655 -0.6658292 -0.7449056 -1.1151520
#>  [2,]  0.4222867 -1.32640195  1.3172479 -1.5087501  1.4217720
#>  [3,]  0.5258783 -1.13269775  0.3337355  0.1315696 -1.2167619
#>  [4,]  0.4992021 -0.71374675  1.4251369 -0.1048885 -1.6879300
#>  [5,]  1.2296179  0.97129225 -0.6668736 -1.2991418 -0.8430539
#>  [6,]  0.4359482  0.11081679 -0.1542000 -1.8107273  1.3052824
#>  [7,] -0.7221102  1.14315208  0.3957588  0.3461719  0.2354969
#>  [8,]  0.8468178 -0.79418025 -0.2890372  0.3031079  0.7766441
#>  [9,]  1.9957190 -0.09344266 -1.0394639  0.6219907 -0.8366753
#> [10,]  0.9254184  0.45330809  0.8981690 -0.7959284 -0.9856772
#> [11,] -0.4709085  0.66014159 -0.5350626  0.0720771  2.2224636
#> [12,]  0.1615685  0.37914099 -1.3957565  1.0514441 -0.1124329
#> attr(,"na.action")
#> [1] 20 17 13 14 15  3 11 12
#> attr(,"class")
#> [1] "omit"
na.omit(X, method = "ie", interp = "before")
#>             [,1]        [,2]       [,3]       [,4]       [,5]
#>  [1,] -1.2900417 -0.43898655 -0.6658292 -0.7449056 -1.1151520
#>  [2,]  0.4222867 -1.32640195  1.3172479 -1.5087501  1.4217720
#>  [3,]  0.5258783 -1.13269775  0.3337355  0.1315696 -1.2167619
#>  [4,]  0.4992021 -0.71374675  1.4251369 -0.1048885 -1.6879300
#>  [5,]  1.2296179  0.97129225 -0.6668736 -1.2991418 -0.8430539
#>  [6,]  0.4359482  0.11081679 -0.1542000 -1.8107273  1.3052824
#>  [7,] -0.7221102  1.14315208  0.3957588  0.3461719  0.2354969
#>  [8,]  0.8468178 -0.79418025 -0.2890372  0.3031079  0.7766441
#>  [9,]  1.9957190 -0.09344266 -1.0394639  0.6219907 -0.8366753
#> [10,]  0.9254184  0.45330809  0.8981690 -0.7959284 -0.9856772
#> [11,] -0.4709085  0.66014159 -0.5350626  0.0720771  2.2224636
#> [12,]  0.1615685  0.37914099 -1.3957565  1.0514441 -0.1124329
#> attr(,"na.action")
#> [1] 20 17 13 14 15  3 11 12
#> attr(,"class")
#> [1] "omit"

## examples with X (which is a matrix, not "timeSeries")
## (these examples are not run automatically as these functions are
## deprecated.) 
if(FALSE){
## Remove Rows with NAs
removeNA(X)
   
## ## subsitute NA's by zeros or column means
## substituteNA(X, type = "zeros")
## substituteNA(X, type = "mean")
##    
## ## interpolate NA's linearily
## interpNA(X, method = "linear")
## # Note the corner missing value cannot be interpolated!
   
## take previous values in a column
interpNA(X, method = "before")
# Also here, the corner value is excluded
}