Dummy time series

Create dummy daily and monthly time series for examples and exploration.

Usage

dummyDailySeries(x = rnorm(365), units = NULL, zone = "", 
                 FinCenter = "")

dummyMonthlySeries(...)

Arguments

x

an object of class timeSeries.

units

an optional character string, which allows to overwrite the current column names of a timeSeries object. By default NULL which means that the column names are selected automatically.

FinCenter

a character with the the location of the financial center named as "continent/city".

zone

the time zone or financial center where the data were recorded.

...

optional arguments passed to timeSeries.

Details

dummyDailySeries creates a timeSeries object with dummy daily dates from a numeric matrix with daily records of unknown dates.

dummyMonthlySeries creates a dummy monthly "timeSeries" object.

Value

a "timeSeries" object

Examples

dummyDailySeries()
#> GMT
#>                      X1
#> 1970-01-01  0.035331347
#> 1970-01-02  0.423148879
#> 1970-01-03  0.202228841
#> 1970-01-04  1.576207457
#> 1970-01-05 -0.469042860
#> 1970-01-06  0.379579908
#> 1970-01-07 -0.534625536
#> 1970-01-08  1.031717303
#> 1970-01-09  0.709566624
#> 1970-01-10  0.078449007
#> 1970-01-11  0.976080435
#> 1970-01-12 -0.262087031
#> 1970-01-13 -1.370296462
#> 1970-01-14  0.196244844
#> 1970-01-15 -1.212355645
#> 1970-01-16 -0.158855728
#> 1970-01-17 -0.668303479
#> 1970-01-18  0.882603162
#> 1970-01-19  0.218294507
#> 1970-01-20  1.765548038
#> 1970-01-21  1.142419102
#> 1970-01-22 -0.591197334
#> 1970-01-23 -0.617830511
#> 1970-01-24 -0.761016948
#> 1970-01-25 -0.287532357
#> 1970-01-26 -2.559374272
#> 1970-01-27 -1.839761220
#> 1970-01-28  1.672603722
#> 1970-01-29 -1.272653368
#> 1970-01-30  0.001369812
#> 1970-01-31 -0.231584632
#> 1970-02-01 -1.947726847
#> 1970-02-02  0.602454282
#> 1970-02-03  0.742540487
#> 1970-02-04  0.234180532
#> 1970-02-05  1.949510250
#> 1970-02-06  0.819855589
#> 1970-02-07 -1.720925329
#> 1970-02-08 -0.338489586
#> 1970-02-09  0.538458280
#> 1970-02-10 -0.467647302
#> 1970-02-11 -1.497852837
#> 1970-02-12 -0.536229820
#> 1970-02-13 -0.907801723
#> 1970-02-14 -0.317032534
#> 1970-02-15 -2.001258679
#> 1970-02-16  0.727027806
#> 1970-02-17  0.940224939
#> 1970-02-18  0.508673317
#> 1970-02-19  0.792654229
#> 1970-02-20  0.143154934
#> 1970-02-21 -0.022807823
#> 1970-02-22 -0.049878707
#> 1970-02-23 -0.927529120
#> 1970-02-24 -0.603169176
#> 1970-02-25 -0.773346887
#> 1970-02-26  0.961407735
#> 1970-02-27 -0.001666551
#> 1970-02-28 -0.227480201
#> 1970-03-01  0.193355496
#> 1970-03-02  0.763771324
#> 1970-03-03  0.166205974
#> 1970-03-04  0.082268895
#> 1970-03-05  1.105133745
#> 1970-03-06  0.411249202
#> 1970-03-07  2.392999665
#> 1970-03-08  1.926076914
#> 1970-03-09 -0.010050865
#> 1970-03-10 -0.230028525
#> 1970-03-11 -0.355881758
#> 1970-03-12  0.558050094
#> 1970-03-13  1.085848515
#> 1970-03-14  0.476502796
#> 1970-03-15 -2.046917682
#> 1970-03-16  0.324064140
#> 1970-03-17  0.255802657
#> 1970-03-18  0.543270491
#> 1970-03-19 -1.099827879
#> 1970-03-20 -1.564810818
#> 1970-03-21 -0.199268086
#> 1970-03-22  0.480072750
#> 1970-03-23  0.180205341
#> 1970-03-24  0.778883392
#> 1970-03-25 -0.860808051
#> 1970-03-26 -0.534062370
#> 1970-03-27 -0.421799090
#> 1970-03-28  1.542860452
#> 1970-03-29 -0.615470704
#> 1970-03-30 -0.390777234
#> 1970-03-31  0.396653298
#> 1970-04-01  0.781569560
#> 1970-04-02 -0.019126050
#> 1970-04-03 -1.944312932
#> 1970-04-04 -0.245506695
#> 1970-04-05  0.352029842
#> 1970-04-06  0.841908204
#> 1970-04-07  0.711083402
#> 1970-04-08 -0.311078313
#> 1970-04-09 -1.035997732
#> 1970-04-10  0.400097235
#> ...
#>  [ reached getRmetricsOption('max.print') | getOption('max.print') -- omitted 265 rows ]]
dummyMonthlySeries()
#> GMT
#>                   TS.1      TS.2
#> 2022-01-01 0.256204173 0.8121176
#> 2022-02-01 0.881703379 0.9931157
#> 2022-03-01 0.816621071 0.1822492
#> 2022-04-01 0.904131538 0.8045545
#> 2022-05-01 0.306723651 0.5473774
#> 2022-06-01 0.470001841 0.5935642
#> 2022-07-01 0.923688679 0.3782178
#> 2022-08-01 0.585954685 0.2984592
#> 2022-09-01 0.002343502 0.5675508
#> 2022-10-01 0.363293023 0.2742155
#> 2022-11-01 0.803557717 0.7835054
#> 2022-12-01 0.468704748 0.5305348