Special Monthly Series

Functions and methods dealing with special monthly "timeSeries" objects.

Usage

countMonthlyRecords(x)

rollMonthlyWindows(x, period = "12m", by = "1m") 
rollMonthlySeries(x, period = "12m", by = "1m", FUN, ...)

Details

countMonthlyRecords computes a "timeSeries" that holds the number of monthly counts of the records.

rollMonthlyWindows computes start and end dates for rolling time windows.

rollMonthlySeries computes the statistics defined by the function FUN over a rolling window, internally computed by the function rollMonthlyWindows. Note, the periods may be overlapping, may be dense, or even may have gaps.

Arguments

x

a "timeSeries" object.

period,by

character strings specifying the rollling period composed by the length of the period and its unit. Examples: "3m" "6m", "12m", and "24m" represent quarterly, semi-annual, annual and bi-annual shifts, respectively. It is the responsibility of the user. to determine proper start of the series

FUN

the function for the statistic to be applied. For example, colMean in the case of aggregation.

...

arguments passed to the function FUN.

Value

for countMonthlyRecords and rollMonthlySeries, a "timeSeries" object.

for rollMonthlyWindows, a list with attribute "control" keeping the start and end dates of the series. The components of the list are:

from

an object from class "timeDate".

to

an object from class "timeDate".

Examples

## Load Microsoft Daily Data Set:
   x <- MSFT 
   
## Count Monthly Records -
   counts <- countMonthlyRecords(x)
   counts
#> GMT
#>            Counts
#> 2000-09-30      3
#> 2000-10-31     22
#> 2000-11-30     21
#> 2000-12-31     20
#> 2001-01-31     21
#> 2001-02-28     19
#> 2001-03-31     22
#> 2001-04-30     20
#> 2001-05-31     22
#> 2001-06-30     21
#> 2001-07-31     21
#> 2001-08-31     23
   
## Quaterly Non-Overlapping Time Periods -
   windows <- rollMonthlyWindows(counts[-1, ], period = "3m", by = "3m") 
   windows
#> $from
#> GMT
#> [1] [2000-10-01] [2000-11-01] [2000-12-01] [2001-01-01] [2001-02-01]
#> [6] [2001-03-01] [2001-04-01] [2001-05-01]
#> 
#> $to
#> GMT
#> [1] [2000-12-31] [2001-01-31] [2001-02-28] [2001-03-31] [2001-04-30]
#> [6] [2001-05-31] [2001-06-30] [2001-07-31]
#> 
#> attr(,"control")
#> GMT
#>        start          end 
#> [2000-10-31] [2001-08-31] 
   
## Nicely Reprint Results as a data.frame -
   data.frame(cbind(FROM=format(windows$from), TO=format(windows$to)))
#>         FROM         TO
#> 1 2000-10-01 2000-12-31
#> 2 2000-11-01 2001-01-31
#> 3 2000-12-01 2001-02-28
#> 4 2001-01-01 2001-03-31
#> 5 2001-02-01 2001-04-30
#> 6 2001-03-01 2001-05-31
#> 7 2001-04-01 2001-06-30
#> 8 2001-05-01 2001-07-31
   
## Compute the average number of monthly trading days per quarter -  
   rollMonthlySeries(counts[-1, ], period = "3m", by = "3m", FUN=mean)
#> GMT
#>              Counts
#> 2000-12-31 21.00000
#> 2001-01-31 20.66667
#> 2001-02-28 20.00000
#> 2001-03-31 20.66667
#> 2001-04-30 20.33333
#> 2001-05-31 21.33333
#> 2001-06-30 21.00000
#> 2001-07-31 21.33333