Handling missing time series values

Functions for handling missing values in 'timeSeries' objects

Usage

# S4 method for timeSeries
na.omit(object, method = c("r", "s", "z", "ir", "iz", "ie"), 
    interp = c("before", "linear", "after"), ...)

## Deprecated:
removeNA(x, ...)
substituteNA(x, type = c("zeros", "mean", "median"), ...)
interpNA(x, method = c("linear", "before", "after"), ...)

Arguments

object

an object of class "timeSeries".

interp, type

Three alternative methods are provided to remove NAs from the data: type="zeros" replaces the missing values with zeros, type="mean" replaces the missing values with the column mean, type="median" replaces the missing values with the column median.

method

for na.omit, the method of handling NAs; for interpNA, how to interpolate the matrix column by column, see Section ‘Details’.

x

a numeric matrix, or any other object which can be transformed into a matrix through x = as.matrix(x, ...). If x is a vector, it will be transformed into a one-dimensional matrix.

...

arguments to be passed to the function as.matrix.

Details

Functions for handling missing values in 'timeSeries' objects and in objects which can be transformed into a vector or a two dimensional matrix.

For na.omit argument method specifies the method how to handle NAs. Can be one of the following strings:

method="s"

na.rm = FALSE, skip, i.e. do nothing,

method="r"

remove NAs,

method="z"

substitute NAs by zeros,

method="ir"

interpolate NAs and remove NAs at the beginning and end of the series,

method="iz"

interpolate NAs and substitute NAs at the beginning and end of the series,

method="ie"

interpolate NAs and extrapolate NAs at the beginning and end of theseries.

For interpNA argument method specifies how to interpolate the matrix column by column. One of the following character strings: "linear", "before", "after". For interpolation the function approx is used.

The functions are listed by topic.

na.omit	Handles NAs,

removeNA	Removes NAs from a matrix object,

substituteNA	substitute NAs by zero, the column mean or median,

interpNA	interpolates NAs using R's "approx" function.

Missing Values in Price and Index Series:

Applied to timeSeries objects the function removeNA just removes rows with NAs from the series. For an interpolation of time series points one can use the function interpNA. Three different methods of interpolation are offered: "linear" does a linear interpolation, "before" uses the previous value, and "after" uses the following value. Note, that the interpolation is done on the index scale and not on the time scale.

Missing Values in Return Series:

For return series the function substituteNA may be useful. The function allows to fill missing values either by method="zeros", the method="mean" or the method="median" value of the appropriate columns.

Note

The functions removeNA, substituteNA and interpNA are older implementations. Please use in all cases if possible the new function na.omit.

When dealing with daily data sets, there exists another function alignDaily Series which can handle missing data in un-aligned calendarical 'timeSeries' objects.

References

Troyanskaya O., Cantor M., Sherlock G., Brown P., Hastie T., Tibshirani R., Botstein D., Altman R.B., (2001); Missing Value Estimation Methods for DNA microarrays Bioinformatics 17, 520--525.

Examples

## Create a Matrix -
   X <- matrix(rnorm(100), ncol = 5)
   
## Replace a Single NA Inside - 
   X[3, 5] <- NA
   
## Replace Three in a Row Inside - 
   X[17, 2:4] <- c(NA, NA, NA)
   
## Replace Three in a Column Inside - 
   X[13:15, 4] <- c(NA, NA, NA)
   
## Replace Two at the Right Border - 
   X[11:12, 5] <- c(NA, NA)
   
## Replace One in the Lower Left Corner - 
   X[20, 1] <- NA
   print(X)
#>              [,1]        [,2]        [,3]        [,4]        [,5]
#>  [1,] -1.38300534  0.36818151 -1.33581811 -1.33646260  0.02155207
#>  [2,] -1.68636263  0.09713429 -0.74909519 -0.04449205 -2.25605388
#>  [3,]  0.99638000  1.48457001 -0.56573440 -0.03258890          NA
#>  [4,] -0.75525465  0.50311697 -0.46460707  0.11172541 -1.04620925
#>  [5,] -0.77794250 -1.06047790  0.23218979 -1.13748485 -1.28139560
#>  [6,]  0.81303352  0.03731678  1.38473259 -0.30385542 -0.70408428
#>  [7,] -0.01840940  0.61124425 -0.69913167  0.90586198  0.61226109
#>  [8,] -0.65331681 -0.52581773  0.58413787  1.39339790 -0.36115428
#>  [9,] -1.07135036  0.78828818 -0.44125019  0.70650028 -0.41463268
#> [10,]  1.44999131  0.32026384 -1.76315270  0.53062526 -0.88296941
#> [11,] -2.15717620  0.10068453 -0.33208134  0.84349377          NA
#> [12,] -0.28623754 -0.71920406 -1.63481270  0.31811098          NA
#> [13,] -0.38423421 -1.20545808 -0.26450379          NA -1.20909152
#> [14,] -0.05014592 -0.29127782  0.29929850          NA -1.54151592
#> [15,]  0.35223126  2.25376438  0.98397636          NA -0.82490739
#> [16,] -1.52088248  0.21382131  0.05082656 -0.48616665 -1.23617668
#> [17,] -0.28058352          NA          NA          NA -2.09464347
#> [18,] -0.02736514  1.19415643  0.54545735  0.09066719 -1.23721039
#> [19,] -2.10717985  1.86686082 -0.63718336  1.54251263  0.27219047
#> [20,]          NA -0.01110177 -0.37886743  0.34644952 -0.97134283
     
## Remove Rows with NAs - 
   removeNA(X)
#>              [,1]        [,2]        [,3]        [,4]        [,5]
#>  [1,] -1.38300534  0.36818151 -1.33581811 -1.33646260  0.02155207
#>  [2,] -1.68636263  0.09713429 -0.74909519 -0.04449205 -2.25605388
#>  [3,] -0.75525465  0.50311697 -0.46460707  0.11172541 -1.04620925
#>  [4,] -0.77794250 -1.06047790  0.23218979 -1.13748485 -1.28139560
#>  [5,]  0.81303352  0.03731678  1.38473259 -0.30385542 -0.70408428
#>  [6,] -0.01840940  0.61124425 -0.69913167  0.90586198  0.61226109
#>  [7,] -0.65331681 -0.52581773  0.58413787  1.39339790 -0.36115428
#>  [8,] -1.07135036  0.78828818 -0.44125019  0.70650028 -0.41463268
#>  [9,]  1.44999131  0.32026384 -1.76315270  0.53062526 -0.88296941
#> [10,] -1.52088248  0.21382131  0.05082656 -0.48616665 -1.23617668
#> [11,] -0.02736514  1.19415643  0.54545735  0.09066719 -1.23721039
#> [12,] -2.10717985  1.86686082 -0.63718336  1.54251263  0.27219047
#> attr(,"na.action")
#> [1] 20 17 13 14 15  3 11 12
#> attr(,"class")
#> [1] "omit"
   
## Subsitute NA's by Zeros or Column Means - 
   substituteNA(X, type = "zeros")
#>              [,1]        [,2]        [,3]        [,4]        [,5]
#>  [1,] -1.38300534  0.36818151 -1.33581811 -1.33646260  0.02155207
#>  [2,] -1.68636263  0.09713429 -0.74909519 -0.04449205 -2.25605388
#>  [3,]  0.99638000  1.48457001 -0.56573440 -0.03258890  0.00000000
#>  [4,] -0.75525465  0.50311697 -0.46460707  0.11172541 -1.04620925
#>  [5,] -0.77794250 -1.06047790  0.23218979 -1.13748485 -1.28139560
#>  [6,]  0.81303352  0.03731678  1.38473259 -0.30385542 -0.70408428
#>  [7,] -0.01840940  0.61124425 -0.69913167  0.90586198  0.61226109
#>  [8,] -0.65331681 -0.52581773  0.58413787  1.39339790 -0.36115428
#>  [9,] -1.07135036  0.78828818 -0.44125019  0.70650028 -0.41463268
#> [10,]  1.44999131  0.32026384 -1.76315270  0.53062526 -0.88296941
#> [11,] -2.15717620  0.10068453 -0.33208134  0.84349377  0.00000000
#> [12,] -0.28623754 -0.71920406 -1.63481270  0.31811098  0.00000000
#> [13,] -0.38423421 -1.20545808 -0.26450379  0.00000000 -1.20909152
#> [14,] -0.05014592 -0.29127782  0.29929850  0.00000000 -1.54151592
#> [15,]  0.35223126  2.25376438  0.98397636  0.00000000 -0.82490739
#> [16,] -1.52088248  0.21382131  0.05082656 -0.48616665 -1.23617668
#> [17,] -0.28058352  0.00000000  0.00000000  0.00000000 -2.09464347
#> [18,] -0.02736514  1.19415643  0.54545735  0.09066719 -1.23721039
#> [19,] -2.10717985  1.86686082 -0.63718336  1.54251263  0.27219047
#> [20,]  0.00000000 -0.01110177 -0.37886743  0.34644952 -0.97134283
   substituteNA(X, type = "mean")
#>              [,1]        [,2]        [,3]        [,4]        [,5]
#>  [1,] -1.38300534  0.36818151 -1.33581811 -1.33646260  0.02155207
#>  [2,] -1.68636263  0.09713429 -0.74909519 -0.04449205 -2.25605388
#>  [3,]  0.99638000  1.48457001 -0.56573440 -0.03258890 -0.89149317
#>  [4,] -0.75525465  0.50311697 -0.46460707  0.11172541 -1.04620925
#>  [5,] -0.77794250 -1.06047790  0.23218979 -1.13748485 -1.28139560
#>  [6,]  0.81303352  0.03731678  1.38473259 -0.30385542 -0.70408428
#>  [7,] -0.01840940  0.61124425 -0.69913167  0.90586198  0.61226109
#>  [8,] -0.65331681 -0.52581773  0.58413787  1.39339790 -0.36115428
#>  [9,] -1.07135036  0.78828818 -0.44125019  0.70650028 -0.41463268
#> [10,]  1.44999131  0.32026384 -1.76315270  0.53062526 -0.88296941
#> [11,] -2.15717620  0.10068453 -0.33208134  0.84349377 -0.89149317
#> [12,] -0.28623754 -0.71920406 -1.63481270  0.31811098 -0.89149317
#> [13,] -0.38423421 -1.20545808 -0.26450379  0.21551840 -1.20909152
#> [14,] -0.05014592 -0.29127782  0.29929850  0.21551840 -1.54151592
#> [15,]  0.35223126  2.25376438  0.98397636  0.21551840 -0.82490739
#> [16,] -1.52088248  0.21382131  0.05082656 -0.48616665 -1.23617668
#> [17,] -0.28058352  0.31716136 -0.27292731  0.21551840 -2.09464347
#> [18,] -0.02736514  1.19415643  0.54545735  0.09066719 -1.23721039
#> [19,] -2.10717985  1.86686082 -0.63718336  1.54251263  0.27219047
#> [20,] -0.50251634 -0.01110177 -0.37886743  0.34644952 -0.97134283
   
## Interpolate NA's Linearily - 
   interpNA(X, method = "linear")
#>              [,1]        [,2]        [,3]        [,4]        [,5]
#>  [1,] -1.38300534  0.36818151 -1.33581811 -1.33646260  0.02155207
#>  [2,] -1.68636263  0.09713429 -0.74909519 -0.04449205 -2.25605388
#>  [3,]  0.99638000  1.48457001 -0.56573440 -0.03258890 -1.65113156
#>  [4,] -0.75525465  0.50311697 -0.46460707  0.11172541 -1.04620925
#>  [5,] -0.77794250 -1.06047790  0.23218979 -1.13748485 -1.28139560
#>  [6,]  0.81303352  0.03731678  1.38473259 -0.30385542 -0.70408428
#>  [7,] -0.01840940  0.61124425 -0.69913167  0.90586198  0.61226109
#>  [8,] -0.65331681 -0.52581773  0.58413787  1.39339790 -0.36115428
#>  [9,] -1.07135036  0.78828818 -0.44125019  0.70650028 -0.41463268
#> [10,]  1.44999131  0.32026384 -1.76315270  0.53062526 -0.88296941
#> [11,] -2.15717620  0.10068453 -0.33208134  0.84349377 -0.99167678
#> [12,] -0.28623754 -0.71920406 -1.63481270  0.31811098 -1.10038415
#> [13,] -0.38423421 -1.20545808 -0.26450379  0.11704157 -1.20909152
#> [14,] -0.05014592 -0.29127782  0.29929850 -0.08402784 -1.54151592
#> [15,]  0.35223126  2.25376438  0.98397636 -0.28509725 -0.82490739
#> [16,] -1.52088248  0.21382131  0.05082656 -0.48616665 -1.23617668
#> [17,] -0.28058352  0.70398887  0.29814195 -0.19774973 -2.09464347
#> [18,] -0.02736514  1.19415643  0.54545735  0.09066719 -1.23721039
#> [19,] -2.10717985  1.86686082 -0.63718336  1.54251263  0.27219047
#> [20,]          NA -0.01110177 -0.37886743  0.34644952 -0.97134283
   # Note the corner missing value cannot be interpolated!
   
## Take Previous Values in a Column - 
   interpNA(X, method = "before")
#>              [,1]        [,2]        [,3]        [,4]        [,5]
#>  [1,] -1.38300534  0.36818151 -1.33581811 -1.33646260  0.02155207
#>  [2,] -1.68636263  0.09713429 -0.74909519 -0.04449205 -2.25605388
#>  [3,]  0.99638000  1.48457001 -0.56573440 -0.03258890 -2.25605388
#>  [4,] -0.75525465  0.50311697 -0.46460707  0.11172541 -1.04620925
#>  [5,] -0.77794250 -1.06047790  0.23218979 -1.13748485 -1.28139560
#>  [6,]  0.81303352  0.03731678  1.38473259 -0.30385542 -0.70408428
#>  [7,] -0.01840940  0.61124425 -0.69913167  0.90586198  0.61226109
#>  [8,] -0.65331681 -0.52581773  0.58413787  1.39339790 -0.36115428
#>  [9,] -1.07135036  0.78828818 -0.44125019  0.70650028 -0.41463268
#> [10,]  1.44999131  0.32026384 -1.76315270  0.53062526 -0.88296941
#> [11,] -2.15717620  0.10068453 -0.33208134  0.84349377 -0.88296941
#> [12,] -0.28623754 -0.71920406 -1.63481270  0.31811098 -0.88296941
#> [13,] -0.38423421 -1.20545808 -0.26450379  0.31811098 -1.20909152
#> [14,] -0.05014592 -0.29127782  0.29929850  0.31811098 -1.54151592
#> [15,]  0.35223126  2.25376438  0.98397636  0.31811098 -0.82490739
#> [16,] -1.52088248  0.21382131  0.05082656 -0.48616665 -1.23617668
#> [17,] -0.28058352  0.21382131  0.05082656 -0.48616665 -2.09464347
#> [18,] -0.02736514  1.19415643  0.54545735  0.09066719 -1.23721039
#> [19,] -2.10717985  1.86686082 -0.63718336  1.54251263  0.27219047
#> [20,]          NA -0.01110177 -0.37886743  0.34644952 -0.97134283
   # Also here, the corner value is excluded