Get and Set Optional Attributes of a 'timeSeries'

Extracts or assigns optional attributes from or to a timeSeries object.

Usage

getAttributes(obj)         
setAttributes(obj) <- value

Arguments

obj

a timeSeries object whose optional attributes are to be accessed.

value

an object, the new value of the attribute, or NULL to remove the attribute.

Details

Each timeSeries object is documented. By default a time series object holds in the documentation slot a string with creation time and the user who has defined it. But this is not all. Optionally the whole creation process and history can be recorded. For this the @documentation slot may have an optional "Attributes" element. This attribute is tracked over the whole life time of the object whenever the time series is changed. Whenever you like to be informed about the optional attributes, or you like to recover them you can dot it, and evenmore, whenever you like to add information as an addiitonal attribute you can also do it.

The two functions getAttributes and setAttributes provide access to and allow to modify the optional attributes of a timeSeries object.

Examples

## Create an artificial timeSeries Object - 
   tS <- dummySeries()
   tS
#> GMT
#>                 TS.1      TS.2
#> 2022-01-01 0.4340961 0.7879249
#> 2022-02-01 0.1724838 0.5952879
#> 2022-03-01 0.2827302 0.5474315
#> 2022-04-01 0.5131803 0.4009878
#> 2022-05-01 0.7534540 0.6252588
#> 2022-06-01 0.8567953 0.4174565
#> 2022-07-01 0.9904504 0.5365323
#> 2022-08-01 0.1511665 0.1585544
#> 2022-09-01 0.3334947 0.6905842
#> 2022-10-01 0.7331391 0.5373382
#> 2022-11-01 0.1273321 0.5663768
#> 2022-12-01 0.6229946 0.3908003

## Get Optional Attributes -
   getAttributes(tS)
#> NULL
   tS@documentation
#> [1] "Sun Oct  9 19:34:55 2022"
   
## Set a new Optional Attribute -
   setAttributes(tS) <- list(what="A dummy Series")
   tS
#> GMT
#>                 TS.1      TS.2
#> 2022-01-01 0.4340961 0.7879249
#> 2022-02-01 0.1724838 0.5952879
#> 2022-03-01 0.2827302 0.5474315
#> 2022-04-01 0.5131803 0.4009878
#> 2022-05-01 0.7534540 0.6252588
#> 2022-06-01 0.8567953 0.4174565
#> 2022-07-01 0.9904504 0.5365323
#> 2022-08-01 0.1511665 0.1585544
#> 2022-09-01 0.3334947 0.6905842
#> 2022-10-01 0.7331391 0.5373382
#> 2022-11-01 0.1273321 0.5663768
#> 2022-12-01 0.6229946 0.3908003
   getAttributes(tS)
#> $what
#> $what[[1]]
#> [1] "A dummy Series"
#> 
#> 
   tS@documentation
#> [1] "Sun Oct  9 19:34:55 2022"
#> attr(,"Attributes")
#> attr(,"Attributes")$what
#> attr(,"Attributes")$what[[1]]
#> [1] "A dummy Series"
#> 
#>