Get and set time stamps of a 'timeSeries'

Functions and methods extracting and modifying positions of 'timeSeries' objects.

Usage

# S4 method for timeSeries
time(x, ...)
# S3 method for timeSeries
time(x) <- value

getTime(x)         
setTime(x) <- value

Arguments

value

a valid value for the time component of x.

x

an object of class timeSeries.

...

optional arguments passed to other methods.

Details

time and time<- are generic functions with methods for class "timeSeries". They get and set the time component of the object.

getTime and setTime are non-generic alternatives are non-generic wrappers of time and time<-, respectively.

Value

for time and getTime, a 'timeDate' object, for time<- and and setTime, the modified 'timeSeries' object.

Examples

## Create Dummy timeSeries -      
   X <- timeSeries(matrix(rnorm(24), 12), timeCalendar())
   
## Return Series Positions -
   getTime(X)
#> GMT
#>  [1] [2022-01-01] [2022-02-01] [2022-03-01] [2022-04-01] [2022-05-01]
#>  [6] [2022-06-01] [2022-07-01] [2022-08-01] [2022-09-01] [2022-10-01]
#> [11] [2022-11-01] [2022-12-01]
   time(X)  
#> GMT
#>  [1] [2022-01-01] [2022-02-01] [2022-03-01] [2022-04-01] [2022-05-01]
#>  [6] [2022-06-01] [2022-07-01] [2022-08-01] [2022-09-01] [2022-10-01]
#> [11] [2022-11-01] [2022-12-01]
  
## Add / Subtract one Day from X
   setTime(X) <- time(X) - 24*3600 # sec
   X
#> GMT
#>                   TS.1        TS.2
#> 2021-12-31 -0.18067150 -0.29008571
#> 2022-01-31  0.23206031  0.08847879
#> 2022-02-28  0.95669554 -1.18344773
#> 2022-03-31 -0.22413890  0.52857993
#> 2022-04-30 -0.11931791 -0.28819707
#> 2022-05-31 -0.43419972 -0.02084050
#> 2022-06-30 -0.51232580 -0.03564996
#> 2022-07-31 -1.67219790  0.66836210
#> 2022-08-31  1.28170459 -0.66726977
#> 2022-09-30  0.07442837 -0.76552161
#> 2022-10-31  1.73175545  0.19676379
#> 2022-11-30 -0.93377821  1.24593943
   time(X) <- time(X) + 24*3600 # sec
   X
#> GMT
#>                   TS.1        TS.2
#> 2022-01-01 -0.18067150 -0.29008571
#> 2022-02-01  0.23206031  0.08847879
#> 2022-03-01  0.95669554 -1.18344773
#> 2022-04-01 -0.22413890  0.52857993
#> 2022-05-01 -0.11931791 -0.28819707
#> 2022-06-01 -0.43419972 -0.02084050
#> 2022-07-01 -0.51232580 -0.03564996
#> 2022-08-01 -1.67219790  0.66836210
#> 2022-09-01  1.28170459 -0.66726977
#> 2022-10-01  0.07442837 -0.76552161
#> 2022-11-01  1.73175545  0.19676379
#> 2022-12-01 -0.93377821  1.24593943